Revert "created blank page and moved detailed questions base to folder"
